[Boot] Added a CMakeLists.txt, started the transition to cmake.
Make is too annoying to deal with at this point, so I decided
to switch to CMake. This feels like the right thing to do, and
it allows for greater portability and easier configuration.
(Although I definitely did have to struggle for a bit to get it to
work with my flat binary NASM bootloader.)