[Kernel] Rewrote VirtualAllocator to use RedBlackTree.
Yay, O(log(N)) allocations and deallocations! There's a catch tho.
As long as each node doesn't know the maximum memory gap on one of its
sides, sized virtual allocations are still O(N) because we have to
potentially iterate the entire allocation tree to find a gap that fits.
So this is something that's on my TODO list. It's not super urgent for
now because this allocator, even as is, is already an insane improvement
from what we had before. :)