[Kernel/Memory] Added BootAllocator.
BootAllocator is a new type of allocator in the system. It allows
anyone to allocate an arbitrary amount of physical pages before the
main MemoryManager is initialized. Right now it's only used to reserve
the kernel image from the physical memory map as well as allocate space
for the initial kernel heap block.