[Kernel/List] Added StandaloneListNode & a List specialization for it.
We now have a List specialization for classes that inherit
StandaloneListNode. Standalone nodes are not owned by the List, which is
very useful for types that are almost exclusively kept inside a list at
all times, e.g Thread/Window in UltraOS. Each time we insert such a node
into a list, we don't have to allocate any space on the heap to store it
since it already has all the members necessary to be inserted. This is
crucial for performance critical code sections, e.g context switching.