[Scripts] KSyms file is now generated from the linker map.
Now using -Map from ld to generate a list of kernel symbols. This allows
us to get rid of the dependency on the ELF format kernel binary. Prior
to this patch we would double-build the kernel each time, once for the
elf and once for the raw binary, for the sole purpose of extracting the
addresses of all kernel symbols. Now that's no longer necessary. Also
added a few new constraints for the symbols, like length and type.
Added stripping of unnecessary kernel:: before each symbol name.