[Kernel/FS] Introduced a basic FAT32 contiguous file layout cache.
This cache aims to reduce the insane overhead of reading a large
file from disk, which requires traversing the entire file allocation
table up to that cluster. We solve the problem by scanning the
table once when a file is first open and generating it's cluster
layout representation in memory by using a sorted array of
contiguous cluster ranges. For a file that's entirely contiguos
on disk, this only takes 8 bytes. In the worse cases, e.g when
dealing with 3+ gigabyte sparse files, it might grow to megabytes
of memory usage for the cache alone. However, for a highly
fragmented disk this seems to be an unavoidable penalty.