[Kernel] Don't use 64 bit atomics in 32 bit kernel.
The way GCC handles 64 bit atomic loads and stores
in 32 bit mode is by using a QWORD FPU load/store
instructions, which is a disaster for a kernel since we
don't preserve the kernel fpu state in any way.
Replace the 64 bit atomic counter with a normal 64
bit counter, and compile with -mgeneral-regs-only.
The way we ensure atomicity is by using a  special
atomic u32 update counter, which is updated every
time we increment the counter, twice.