[Kernel/Interrupts] Began work on DeferredIRQManager/Handler.
This will be used for invoking irq handlers in a deferred manner,
which is very useful for handlers that might have to do a lot
of work including memory allocation or waiting for something,
like xHCI and reacting to port status change events. Such handlers
will simply be executed later in a normal ring 0 thread context by the
DeferredIRQManager.